chore(organization): restructure organization
