+++ next-tavla/next.config.js
@@ -8,7 +8,7 @@ const nextConfig = {
         remotePatterns: [
             {
                 protocol: 'https',
-                hostname: 'firebasestorage.googleapis.com',
+                hostname: 'storage.googleapis.com',
             },
         ],
     },
