+++ next-tavla/src/Admin/scenarios/Organization/styles.module.css
@@ -1,0 +1,5 @@
+.organization {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 2em;
+}
